Adding support for transaction commands. No support for the actual transaction will be added on this version.
The Redis transactions require multi bulk sending of commands, in a
serial way. If you send commands and in the middle another client sends
them, then you loose the transaction.
So this requires a lazy sending like approach to figure out how to send
stuff. And I will not implement it on this version of the library. Only
in feature ones. Please read the roadmap and feature lists for further
information on this subject